Update AttrRadioButtonList.js
When the widget is on a popup form the subscription to the changes is fired although the form and widget are destroyed. Added an extra check. The behavior is that the popup can not be closed.
It looks like a timing issue. The widget should no longer receive updates. Tested in Mx 5.13.1